edwards25519: reorder code a bit
